+++ lib/versions/plugins.rb
@@ -1,4 +1,4 @@
-# Copyright 2018 Lars Eric Scheidler
+# Copyright 2019 Lars Eric Scheidler
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,6 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-module Versions
-  VERSION = "0.2.0"
-end
+require_relative "plugins/docker"
+require_relative "plugins/release_directory"