+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.2.0 (2019-11-01)
+==================
+
+- introduce plugins for following deployment types:
+  - release directory (current, previous symlink)
+  - docker image
+
 0.1.0 (2019-08-27)
 ==================